Split into 3 files
* webview_flutter_tizen.dart : webview part
* platform_view.dart : platform view part imported without modification
* platform_view_tizen.dart : custom platform view part
